refactor(button): refactor the logic and design of the button
